refactor: Split service github out of core
